增加数据购买功能

+++ src/main/java/edu/nyu/dss/similarity/index/DatasetPriceMap.java
@@ -2,8 +2,9 @@ package edu.nyu.dss.similarity.index;
 
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.util.HashMap;
 
 @Component
-public class DatasetPriceMap extends HashMap<Integer, Double> {
+public class DatasetPriceMap extends HashMap<Integer, BigDecimal> {
 }
